Cleanup trace SDK
