--- pkg/internal/export/otel/traces_receiver_test.go
@@ -1,288 +1,0 @@
-package otel
-
-import (
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/ptrace"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
-
-	"github.com/grafana/beyla/pkg/internal/request"
-)
-
-func TestGenerateTraces(t *testing.T) {
-	t.Run("test with subtraces - with parent spanId", func(t *testing.T) {
-		start := time.Now()
-		parentSpanID, _ := trace.SpanIDFromHex("89cbc1f60aab3b04")
-		spanID, _ := trace.SpanIDFromHex("89cbc1f60aab3b01")
-		traceID, _ := trace.TraceIDFromHex("eae56fbbec9505c102e8aabfc6b5c481")
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			Start:        start.Add(time.Second).UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-			Status:       200,
-			ParentSpanID: parentSpanID,
-			TraceID:      traceID,
-			SpanID:       spanID,
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 3, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-		assert.Equal(t, "in queue", spans.At(0).Name())
-		assert.Equal(t, "processing", spans.At(1).Name())
-		assert.Equal(t, "GET /test", spans.At(2).Name())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(0).Kind())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(1).Kind())
-		assert.Equal(t, ptrace.SpanKindServer, spans.At(2).Kind())
-
-		assert.NotEmpty(t, spans.At(2).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(2).TraceID().String())
-		topSpanID := spans.At(2).SpanID().String()
-		assert.Equal(t, parentSpanID.String(), spans.At(2).ParentSpanID().String())
-
-		assert.NotEmpty(t, spans.At(0).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(0).TraceID().String())
-		assert.Equal(t, topSpanID, spans.At(0).ParentSpanID().String())
-
-		assert.Equal(t, spanID.String(), spans.At(1).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(1).TraceID().String())
-		assert.Equal(t, topSpanID, spans.At(1).ParentSpanID().String())
-
-		assert.NotEqual(t, spans.At(0).SpanID().String(), spans.At(1).SpanID().String())
-		assert.NotEqual(t, spans.At(1).SpanID().String(), spans.At(2).SpanID().String())
-	})
-
-	t.Run("test with subtraces - ids set bpf layer", func(t *testing.T) {
-		start := time.Now()
-		spanID, _ := trace.SpanIDFromHex("89cbc1f60aab3b04")
-		traceID, _ := trace.TraceIDFromHex("eae56fbbec9505c102e8aabfc6b5c481")
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			Start:        start.Add(time.Second).UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-			Status:       200,
-			SpanID:       spanID,
-			TraceID:      traceID,
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 3, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-		assert.Equal(t, "in queue", spans.At(0).Name())
-		assert.Equal(t, "processing", spans.At(1).Name())
-		assert.Equal(t, "GET /test", spans.At(2).Name())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(0).Kind())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(1).Kind())
-		assert.Equal(t, ptrace.SpanKindServer, spans.At(2).Kind())
-
-		assert.NotEmpty(t, spans.At(0).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(0).TraceID().String())
-
-		assert.Equal(t, spanID.String(), spans.At(1).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(1).TraceID().String())
-
-		assert.NotEmpty(t, spans.At(2).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(2).TraceID().String())
-		assert.NotEqual(t, spans.At(0).SpanID().String(), spans.At(1).SpanID().String())
-		assert.NotEqual(t, spans.At(1).SpanID().String(), spans.At(2).SpanID().String())
-	})
-
-	t.Run("test with subtraces - generated ids", func(t *testing.T) {
-		start := time.Now()
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			Start:        start.Add(time.Second).UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-			Status:       200,
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 3, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-		assert.Equal(t, "in queue", spans.At(0).Name())
-		assert.Equal(t, "processing", spans.At(1).Name())
-		assert.Equal(t, "GET /test", spans.At(2).Name())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(0).Kind())
-		assert.Equal(t, ptrace.SpanKindInternal, spans.At(1).Kind())
-		assert.Equal(t, ptrace.SpanKindServer, spans.At(2).Kind())
-
-		assert.NotEmpty(t, spans.At(0).SpanID().String())
-		assert.NotEmpty(t, spans.At(0).TraceID().String())
-		assert.NotEmpty(t, spans.At(1).SpanID().String())
-		assert.NotEmpty(t, spans.At(1).TraceID().String())
-		assert.NotEmpty(t, spans.At(2).SpanID().String())
-		assert.NotEmpty(t, spans.At(2).TraceID().String())
-		assert.NotEqual(t, spans.At(0).SpanID().String(), spans.At(1).SpanID().String())
-		assert.NotEqual(t, spans.At(1).SpanID().String(), spans.At(2).SpanID().String())
-	})
-
-	t.Run("test without subspans - ids set bpf layer", func(t *testing.T) {
-		start := time.Now()
-		spanID, _ := trace.SpanIDFromHex("89cbc1f60aab3b04")
-		traceID, _ := trace.TraceIDFromHex("eae56fbbec9505c102e8aabfc6b5c481")
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-			Status:       200,
-			SpanID:       spanID,
-			TraceID:      traceID,
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-
-		assert.Equal(t, spanID.String(), spans.At(0).SpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(0).TraceID().String())
-	})
-
-	t.Run("test without subspans - with parent spanId", func(t *testing.T) {
-		start := time.Now()
-		parentSpanID, _ := trace.SpanIDFromHex("89cbc1f60aab3b04")
-		traceID, _ := trace.TraceIDFromHex("eae56fbbec9505c102e8aabfc6b5c481")
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-			Status:       200,
-			ParentSpanID: parentSpanID,
-			TraceID:      traceID,
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-
-		assert.Equal(t, parentSpanID.String(), spans.At(0).ParentSpanID().String())
-		assert.Equal(t, traceID.String(), spans.At(0).TraceID().String())
-	})
-
-	t.Run("test without subspans - generated ids", func(t *testing.T) {
-		start := time.Now()
-		span := &request.Span{
-			Type:         request.EventTypeHTTP,
-			RequestStart: start.UnixNano(),
-			End:          start.Add(3 * time.Second).UnixNano(),
-			Method:       "GET",
-			Route:        "/test",
-		}
-		traces := GenerateTraces(span)
-
-		assert.Equal(t, 1, traces.ResourceSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().Len())
-		assert.Equal(t, 1, traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans().Len())
-		spans := traces.ResourceSpans().At(0).ScopeSpans().At(0).Spans()
-
-		assert.NotEmpty(t, spans.At(0).SpanID().String())
-		assert.NotEmpty(t, spans.At(0).TraceID().String())
-	})
-}
-
-func TestAttrsToMap(t *testing.T) {
-	t.Run("test with string attribute", func(t *testing.T) {
-		attrs := []attribute.KeyValue{
-			attribute.String("key1", "value1"),
-			attribute.String("key2", "value2"),
-		}
-		expected := pcommon.NewMap()
-		expected.PutStr("key1", "value1")
-		expected.PutStr("key2", "value2")
-
-		result := attrsToMap(attrs)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("test with int attribute", func(t *testing.T) {
-		attrs := []attribute.KeyValue{
-			attribute.Int64("key1", 10),
-			attribute.Int64("key2", 20),
-		}
-		expected := pcommon.NewMap()
-		expected.PutInt("key1", 10)
-		expected.PutInt("key2", 20)
-
-		result := attrsToMap(attrs)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("test with float attribute", func(t *testing.T) {
-		attrs := []attribute.KeyValue{
-			attribute.Float64("key1", 3.14),
-			attribute.Float64("key2", 2.718),
-		}
-		expected := pcommon.NewMap()
-		expected.PutDouble("key1", 3.14)
-		expected.PutDouble("key2", 2.718)
-
-		result := attrsToMap(attrs)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("test with bool attribute", func(t *testing.T) {
-		attrs := []attribute.KeyValue{
-			attribute.Bool("key1", true),
-			attribute.Bool("key2", false),
-		}
-		expected := pcommon.NewMap()
-		expected.PutBool("key1", true)
-		expected.PutBool("key2", false)
-
-		result := attrsToMap(attrs)
-		assert.Equal(t, expected, result)
-	})
-}
-
-func TestCodeToStatusCode(t *testing.T) {
-	t.Run("test with unset code", func(t *testing.T) {
-		code := codes.Unset
-		expected := ptrace.StatusCodeUnset
-
-		result := codeToStatusCode(code)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("test with error code", func(t *testing.T) {
-		code := codes.Error
-		expected := ptrace.StatusCodeError
-
-		result := codeToStatusCode(code)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("test with ok code", func(t *testing.T) {
-		code := codes.Ok
-		expected := ptrace.StatusCodeOk
-
-		result := codeToStatusCode(code)
-		assert.Equal(t, expected, result)
-	})
-}
